refactor: extract CashGameFields and TournamentFields from SessionForm
Split session-form.tsx into three focused components:
- CashGameFields: buy-in/cash-out, variant, blinds, ante, table size
- TournamentFields: buy-in, entry fee, placement, entries, prize, rebuys, addon, bounty
- SessionForm: shared layout (type toggle, date, time, tags, memo) + delegates to above

Also extracts parseCashGameFields/parseTournamentFields for form data parsing.
